requirements.txt
id bot= https://t.me/timer_chat_bot

annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
openai==1.10.0
pydantic==2.6.0
pydantic_core==2.16.1
python-telegram-bot==20.7
sniffio==1.3.0
telegram_text==0.2.0
tqdm==4.66.1
typing_extensions==4.9.0
